+++ example/mnist/test_api.py
@@ -7,7 +7,6 @@ import requests
 import numpy as np
 import json
 
-
 def test_api(image_hex, host, port):
     url = 'http://0.0.0.0:8181/v1/predict/'
     js = {"image": image_hex}