+++ thrift4DL/thrift_file/gen-py/__init__.py
@@ -3,5 +3,3 @@
 # This software is released under the MIT License.
 # https://opensource.org/licenses/MIT
 
-from .server import TModelPoolServer
-from .handler import VisionHandler