+++ thrift4DL/thrift_file/gen-py/thrift4DL/constants.py
@@ -11,8 +11,4 @@ from thrift.protocol.TProtocol import TProtocolException
 from thrift.TRecursive import fix_spec
 
 import sys
-
-from thrift.transport import TTransport
-all_structs = []
-fix_spec(all_structs)
-del all_structs
+from .ttypes import *